fix type safety issue for SunPro toolchain
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@9640 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ repro/CommandServer.cxx
@@ -211,7 +211,7 @@ CommandServer::handleGetDnsCacheRequest(unsigned int connectionId, unsigned int 
 {
    InfoLog(<< "CommandServer::handleGetDnsCacheRequest");
 
-   mReproRunner.getProxy()->getStack().getDnsCacheDump(make_pair(connectionId, requestId), this);
+   mReproRunner.getProxy()->getStack().getDnsCacheDump(make_pair((unsigned long)connectionId, (unsigned long)requestId), this);
    // Note: Response will be sent when callback is invoked
 }
